+++ src/utils/distance_util.h
@@ -28,7 +28,7 @@ std::vector<std::vector<int64_t>> CalcDistances(
     const std::vector<std::vector<int64_t>>& locations,
     DistanceAlgorithm *algorithm);
 
-int64_t CalcDistanceCost(const std::vector<int64_t>& nodes_permutation, const std::vector<std::vector<int64_t>>& distances);
+int64_t CalcDistanceCost(const std::vector<int>& nodes_permutation, const std::vector<std::vector<int64_t>>& distances);
 } // namespace covid19
 
 #endif // COVID19_DISTANCE_H_
