+++ src/utils/file_util.h
@@ -10,7 +10,7 @@ namespace covid19
 {
 struct ChristofidesDataModel
 {
-    int capacity = 1;
+    int64_t capacity = 1;
     std::vector<std::vector<int64_t>> nodes{};
 };
